# pkg/api/openapi.yaml
openapi: 3.0.0
info:
  title: Defeed API
  version: 1.0.0
  description: Defeat your fragmented feeds 💪
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.defeed.co
    description: Production server
paths:
...
  /sources:
    get:
      summary: List available sources
      operationId: listSources
      tags:
        - sources
      parameters:
        - name: query
          in: query
          description: Filter sources by name or description.
          schema:
            type: string
        - name: topics
          in: query
          description: "Optional list of user interests to personalize results. Example: interests=llms&interests=startups"
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/TopicTag'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Source'
  /sources/{uid}:
    get:
      summary: Get source by UID
      operationId: getSource
      tags:
        - sources
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Source info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          description: Source not found
  /feeds:
    post:
      summary: Create a feed belonging to the authenticated user
      operationId: createOwnFeed
      tags:
        - feeds
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedRequest"
      responses:
        '200':
          description: Feed created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        '401':
          description: Unauthorized - Invalid or missing authentication token
    get:
      summary: List public feeds and/or those belonging to the authenticated user
      operationId: listFeeds
      tags:
        - feeds
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of feeds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feed"
        '401':
          description: Unauthorized - Invalid or missing authentication token
  /feeds/{uid}:
    put:
      summary: Update a feed belonging to the authenticated user
      operationId: updateOwnFeed
      tags:
        - feeds
      security:
        - bearerAuth: []
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedRequest"
      responses:
        '200':
          description: Feed updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        '401':
          description: Unauthorized - Invalid or missing authentication token
        '404':
          description: Feed not found
    delete:
      summary: Delete a feed belonging to the authenticated user
      operationId: deleteOwnFeed
      tags:
        - feeds
      security:
        - bearerAuth: []
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Feed deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '401':
          description: Unauthorized - Invalid or missing authentication token
        '404':
          description: Feed not found
  /feeds/{uid}/activities:
    get:
      summary: List activities for a feed
      operationId: listFeedActivities
      tags:
        - feeds
      security:
        - bearerAuth: []
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          description: Time period to filter activities from. Defaults to 'all' for all time.
          schema:
            $ref: '#/components/schemas/ActivityPeriod'
        - name: sortBy
          in: query
          description: Sort method.
          schema:
            $ref: '#/components/schemas/ActivitySortBy'
        - name: query
          in: query
          description: Filter query. Authenticated users can override the default feed query.
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of activities to return.
          schema:
            type: integer
            default: 20
        - name: rewriteQuery
          in: query
          description: Whether to rewrite the query to sub-queries and return results by topics.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Activities list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivitiesListResponse'
        '401':
          description: Unauthorized - Invalid or missing authentication token
        '404':
          description: Feed not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      in: header
      name: Authorization
      bearerFormat: JWT
      description: Bearer token authentication. Include the token in the Authorization header as 'Bearer {token}'.
  schemas:
...
    UpdateFeedRequest:
      allOf:
        - $ref: '#/components/schemas/CreateFeedRequest'
    CreateFeedRequest:
      type: object
      required:
        - name
        - icon
        - query
        - sourceUids
      properties:
        name:
          type: string
        icon:
          type: string
        query:
          type: string
        sourceUids:
          type: array
          items:
            type: string
    Feed:
      type: object
      required:
        - uid
        - name
        - icon
        - query
        - sourceUids
        - isPublic
        - createdBy
        - createdAt
      properties:
        uid:
          type: string
        name:
          type: string
        icon:
          type: string
        query:
          type: string
        sourceUids:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        createdBy:
          description: "ID of the user who created and owns the feed. Feed can only be modified by him."
          type: string
        createdAt:
          type: string
          format: date-time
    Source:
      type: object
      required:
        - uid
        - type
        - name
        - description
        - url
        - iconUrl
        - topicTags
      properties:
        uid:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        name:
          type: string
        description:
          type: string
        url:
          type: string
        iconUrl:
          type: string
        topicTags:
          type: array
          items:
            $ref: '#/components/schemas/TopicTag'
    TopicTag:
      type: string
      description: Specific niche technology/startup interests
      enum:
        - large_language_models
        - startups
        - devtools
        - web_performance
        - distributed_systems
        - databases
        - security_engineering
        - systems_programming
        - product_management
        - growth_engineering
        - artificial_intelligence
        - robotics
        - open_source
        - cloud_infrastructure
        - computer_science
        - science
    SourceType:
      type: string
      enum:
        - mastodonAccount
        - mastodonTag
        - hackernewsPosts
        - redditSubreddit
        - lobstersTag
        - lobstersFeed
        - rssFeed
        - githubReleases
        - githubIssues
        - githubTopics
        - changedetectionWebsite
        - productHuntPosts
    ActivitySortBy:
      type: string
      enum:
        - similarity
        - creationDate
      default: similarity
    ActivityPeriod:
      type: string
      enum:
        - all
        - month
        - week
        - day
      default: all
      description: Time period to filter activities from. 'month' means last month, 'week' means last week, 'day' means last day.
      x-go-enum:
        - all
        - month
        - week
        - day
...
    ActivitiesListResponse:
      type: object
      required:
        - topics
        - results
      properties:
        topics:
          type: array
          items:
            $ref: '#/components/schemas/ActivityTopic'
        nextCursor:
          type: string
          description: Cursor to use for fetching the next page of results
        hasMore:
          type: boolean
          description: Whether there are more results available
        results:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
    ActivityTopic:
      type: object
      required:
        - title
        - emoji
        - summary
        - queries
        - activityIds
      properties:
        title:
          type: string
          description: Title of the topic.
        emoji:
          type: string
          description: Emoji character for the topic.
        summary:
          type: string
          description: Summary of the activities in this topic.
        queries:
          type: array
          description: LLM generated sub-queries used to filter activities for this topic.
          items:
            type: string
        activityIds:
          type: array
          items:
            type: string
          description: List of activity IDs in this topic.
    Activity:
      type: object
      required:
        - uid
        - sourceUids
        - sourceType
        - title
        - shortSummary
        - fullSummary
        - body
        - url
        - imageUrl
        - createdAt
        - upvotesCount
        - commentsCount
        - amplificationCount
      properties:
        uid:
          type: string
        sourceUids:
          type: array
          items:
            type: string
        sourceType:
          $ref: '#/components/schemas/SourceType'
        title:
          type: string
        shortSummary:
          type: string
          description: One-line short plain text summary.
        fullSummary:
          type: string
          description: One-paragraph markdown summary.
        body:
          type: string
        url:
          type: string
          format: url
        imageUrl:
          type: string
          format: url
        createdAt:
          type: string
          format: date-time
        similarity:
          type: number
          format: float
        upvotesCount:
          type: integer
          description: Number of upvotes/likes. -1 if not available.
        commentsCount:
          type: integer
          description: Number of comments/discussions. -1 if not available.
        amplificationCount:
          type: integer
          description: Number of shares/reposts/forks/etc. -1 if not available.